+++ clusters/k3d/dependencies/k3d/deploy_k3d.sh
@@ -1,0 +1,7 @@
+#!/usr/bin/env bash
+
+set -e
+trap 'echo ❌ exit at ${0}:${LINENO}, command was: ${BASH_COMMAND} 1>&2' ERR
+set -x
+
+k3d cluster create ${CI_JOB_ID} --config ${PIPELINE_REPO_DESTINATION}/clusters/k3d/dependencies/k3d/config.yaml --network ${CI_JOB_ID}