+++ infrastructure/aws/dependencies/terraform/env/dev/dev.tf
@@ -1,0 +1,5 @@
+module "dev" {
+  source   = "../../main"
+  env      = "dev"
+  vpc_cidr = "10.255.0.0/16"
+}